Handle overlapping ranges (#3)
Previously we implicitly assumed that character ranges in a NFA/DFA state
cannot overlap, so for example, I cannot have rules `'0'-'9'` and `'5'-'9'` in
the NFA/DFA state. This is obviously incorrect, as reported in #2.

This patch introduces a "range map" that maps `u32` ranges to values. A range
can have multiple values (used in NFA-to-DFA conversion). When two ranges
overlap, the overlapping part has union of values of the overlapping ranges.

Using this type instead of `HashMap<(char, char), HashSet<NfaStateIdx>>`
automatically fixes the issue, as the handling of overlapping ranges
effectively implements switching to states of all of the overlapping ranges in
the NFA. So most of the code is just implementing the `RangeMap` type. Rest are
tests and replacing hash maps with `RangeMap`.

I also benchmarked this patch. For the Lua 5.1 lexer, this PR reduces compile
times 9%. Generated code is identical except orders of alternatives in or
patterns for ranges.

Fixes #2