Inline states with single predecessors
If a state has only one predecessor we now "inline" it in the
predecessor state. This reduces number of states in the generated code
and redundant looping.

Criterion output:

    Lex Lua files time:   [2.5828 ms 2.6028 ms 2.6235 ms]
                  change: [-6.1421% -5.3836% -4.7229%] (p = 0.00 < 0.05)
                  Performance has improved.

Code size of Lua 5.1 lexer reduced from 9334 to 7072 loc (-24.2%).

The code is an absolute mess and codegen became even more messy, but
that's fine, we can refactor it later when I'm back at my desktop.