Refactor range map implementation:
Current implementation of range map needs ranges for the parts of the
space without any values. For example, empty range map is represented as

    [(MIN, MAX, [])]

This is mainly to keep `insert_values` implementation simple.

However, it causes problems when in a few other places. In DFA, a range
map can map a range to only one value. Currently this invariant is
maintained in runtime, with assertions like

    assert_eq!(values.len(), 1);

To maintain this invariant in compile time, we need to make `RangeMap`
parametric on the mapped value, rather than vector elements (i.e. each
range is mapped to a vector, the map type is parametric on vector
elements).

This is possible with the current implementation with types like this:

    type NfaRangeMap = RangeMap<Vec<NfaStateIdx>>;
    type DfaRangeMap = RangeMap<Option<DfaStateIdx>>;

The `Option` in `DfaRangeMap` is needed to allow representing "unmapped
space" in the map.

With this refactoring, we will be able to do:

    type NfaRangeMap = RangeMap<NonEmptyVec<NfaStateIdx>>;
        // or even better: NonEmptySet

    type DfaRangeMap = RangeMap<DfaStateIdx>;

This refactoring will be done in subsequent commit(s).

Iteration of RangeMap also becomes simpler as we don't have to filter
out empty space.