Refactor RangeMap:
RangeMap is now parametric on the value type. Previously the value was
always a `Vec`, and the type parameter was for the `Vec` elements. This
caused problems in a few places:

- In NFA, we want to map ranges to sets, rather than vectors, so we
  couldn't use RangeMap in NFA.

- In DFA, we want to map a range to one value. Previously we had runtime
  checks in use sites to make sure this is the case. Now we use
  `StateIdx` instead of `Vec<StateIdx>` for ranges in DFAs, and check in
  construction time that we map a range to only one `StateIdx`.

This sets the stage for #24. Now that we use `RangeMap` everywhere where
we deal with character sets, exclusion can be implemented easily once
we implement a `RangeMap::remove` method.